+++ test/dygraph_to_static/test_deepcopy.py
@@ -14,6 +14,7 @@
 
 import unittest
 from copy import deepcopy
+from types import MethodType
 
 import numpy as np
 from dygraph_to_static_utils import Dy2StTestBase, test_ast_only
@@ -89,7 +90,7 @@ class TestDeepCopy(Dy2StTestBase):
         out = model(x)
 
         copied_model = deepcopy(static_model)
-        # self.assertIsInstance(copied_model.inner.forward, WeakMethod)
+        self.assertIsInstance(copied_model.inner.forward, MethodType)
         self.assertIsNot(static_model.inner.forward, copied_model.inner.forward)
         self.assertIsNot(
             static_model.inner.forward.__self__,
@@ -101,7 +102,7 @@ class TestDeepCopy(Dy2StTestBase):
         copied_out = copied_model(x)
 
         copied_model.forward.rollback()
-        # self.assertFalse(isinstance(copied_model.inner.forward, WeakMethod))
+        self.assertIsInstance(copied_model.inner.forward, MethodType)
         copied_model(x)
         copied_rollback_out = copied_model(x)
         np.testing.assert_array_equal(out.numpy(), copied_out.numpy())
